fix: rename wyrand module for less ambiguity
